Fix comments pages
