+++ package.json
@@ -10,30 +10,30 @@
     "export": "next export"
   },
   "dependencies": {
-    "antd": "^4.24.1",
+    "antd": "^5.2.0",
     "date-fns": "^2.29.3",
     "draft-js": "^0.11.7",
     "gray-matter": "^4.0.3",
-    "mongoose": "^6.7.2",
-    "next": "13.0.3",
+    "mongoose": "^6.9.1",
+    "next": "13.1.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "rehype-raw": "^6.1.1",
     "rehype-stringify": "^9.0.3",
     "remark-parse": "^10.0.1",
     "remark-rehype": "^10.1.0",
-    "swr": "^1.3.0",
+    "swr": "^2.0.3",
     "unified": "^10.1.2"
   },
   "devDependencies": {
     "@playwright/test": "^1.30.0",
-    "@types/draft-js": "^0.11.9",
-    "@types/node": "18.11.9",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.8",
-    "eslint": "8.27.0",
-    "eslint-config-next": "13.0.3",
+    "@types/draft-js": "^0.11.10",
+    "@types/node": "18.13.0",
+    "@types/react": "18.0.28",
+    "@types/react-dom": "18.0.10",
+    "eslint": "8.34.0",
+    "eslint-config-next": "13.1.6",
     "playwright": "^1.30.0",
-    "typescript": "4.8.4"
+    "typescript": "4.9.5"
   }
 }
