+++ pages/api/comments/index.js
@@ -9,7 +9,7 @@ export default async function handler(req, res) {
   switch (method) {
     case 'GET':
       try {
-        const pets = await Comments.find({}) /* find all the data in our database */
+        const pets = await Comments.find({}).limit(25) /* find all the data in our database */
         res.status(200).json({ success: true, data: pets })
       } catch (error) {
         res.status(400).json({ success: false })
